feat: add conditional to fk error

+++ src/data/services/Agency/DeleteAgency.service.js
@@ -1,8 +1,12 @@
-import { AgencyRepository } from '../../database/repositories';
+import {
+  AgencyRepository,
+  ProjectRepository,
+} from '../../database/repositories';
 
 export class DeleteAgencyService {
   async execute({ id }) {
     const repository = new AgencyRepository();
+    const projectRepository = new ProjectRepository();
 
     const verifyAgencyExists = await repository.findAgencyById({
       id,
@@ -12,6 +16,17 @@ export class DeleteAgencyService {
       return {
         error: `Não há nenhum orgão registrado com este ID -> ${id}.`,
       };
+
+    const verifyFk = await projectRepository.verifyRelationAgency({
+      id_agency: id,
+    });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir o Orgão pois existem Projetos associados.',
+      };
+    }
 
     await repository.deleteAgency({
       id,
