+++ src/data/services/Category/DeleteCategory.service.js
@@ -1,8 +1,12 @@
-import { CategoryRepository } from '../../database/repositories';
+import {
+  CategoryRepository,
+  ProjectRepository,
+} from '../../database/repositories';
 
 export class DeleteCategoryService {
   async execute({ id }) {
     const repository = new CategoryRepository();
+    const projectRepository = new ProjectRepository();
 
     const verifyCategoryExists = await repository.findCategoryById({
       id,
@@ -10,6 +14,17 @@ export class DeleteCategoryService {
 
     if (!verifyCategoryExists)
       return { error: `Não existe uma categoria com este ID -> ${id}.` };
+
+    const verifyFk = await projectRepository.verifyRelationCategory({
+      id_category: id,
+    });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir a Categoria pois existem Projetos associados.',
+      };
+    }
 
     await repository.deleteCategory({
       id,
